backend/pyproject.toml
[project]
name = "aztec-assess-backend"
version = "0.1.0"
description = "Backend for the Aztec Assess web application"
authors = [
    {name = "talberry",email = "52144102+talberry@users.noreply.github.com"}
]
license = "MIT"
readme = "../README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.12"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.9.0"
psycopg2-binary = "^2.9.11"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.5.1"
dj-database-url = "^3.1.2"
drf-spectacular = "^0.29.0"
drf-spectacular-sidecar = "^2026.3.1"
requests = "^2.33.0"
google-cloud-storage = ">=2.14,<3"
google-generativeai = ">=0.8"
numpy = "^2.4.3"
pypdf = "^6.9.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.6"
mypy = "^1.19.1"
pre-commit = "^4.5.1"
django-stubs = "^5.2.9"
djangorestframework-stubs = "^3.16.8"
types-requests = "^2.32.0"

[tool.poetry.group.test.dependencies]
factory-boy = "^3.3.3"
coverage = "^7.13.4"
pytest = "^9.0.2"
pytest-django = "^4.12.0"
pytest-cov = "^7.0.0"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "DJ", # flake8-django
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "DJ01",  # Django model requires __str__ method
    "DJ008",  # False positive: model __str__ present
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"*/migrations/*" = ["ALL"]  # Ignore all rules in migrations

[tool.ruff.lint.isort]
known-first-party = ["adaptive_testing", "apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "adaptive_testing.settings.testing"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests"
]

[tool.coverage.run]
source = ["adaptive_testing", "apps"]
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/tests/*",
    "*/settings/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
disable_error_code = ["var-annotated"]

[[tool.mypy.overrides]]
module = "django.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "decouple.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dj_database_url.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "apps.*.tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "apps.accounts.models"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "apps.quizzes.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "apps.quizzes.tests.*"
disallow_untyped_defs = false
disable_error_code = ["attr-defined", "union-attr"]

[[tool.mypy.overrides]]
module = "apps.ai.tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "google.generativeai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.generativeai"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "adaptive_testing.settings"